fix(web): separate model and effort selectors
