+++ client/src/__tests__/mainchat-model-selector.test.ts
@@ -4,7 +4,7 @@ import { mount } from "@vue/test-utils";
 import type { ModelConfig } from "../api/types";
 
 import MainChat from "../components/MainChat.vue";
-import MainChatModelPopover from "../components/MainChatModelPopover.vue";
+import MainChatModelSelectors from "../components/MainChatModelSelectors.vue";
 
 function makeModel(id: string, displayName: string, provider: string): ModelConfig {
   return {
@@ -25,15 +25,15 @@ describe("MainChat model selector", () => {
     busy: false,
   } as const;
 
-  const popoverBaseProps = {
+  const selectorBaseProps = {
     connected: true,
     busy: false,
   } as const;
 
   it("renders the model label without an agent selector or appended id", () => {
-    const wrapper = mount(MainChatModelPopover, {
+    const wrapper = mount(MainChatModelSelectors, {
       props: {
-        ...popoverBaseProps,
+        ...selectorBaseProps,
         agents: [{ id: "codex", name: "Codex", ready: true }],
         activeAgentId: "codex",
         models: [makeModel("gpt-4.1", "GPT-4.1", "openai")],
@@ -44,7 +44,7 @@ describe("MainChat model selector", () => {
     const agentSelect = wrapper.find('select[aria-label="Select agent"]');
     expect(agentSelect.exists()).toBe(false);
 
-    const modelSelect = wrapper.find('[data-testid="chat-model-popover-toggle"]');
+    const modelSelect = wrapper.find('[data-testid="chat-model-select"]');
     expect(modelSelect.exists()).toBe(true);
     expect(modelSelect.text()).toContain("GPT-4.1");
     expect(modelSelect.text()).not.toContain("(gpt-4.1)");
@@ -61,9 +61,9 @@ describe("MainChat model selector", () => {
       allowedAgents: ["custom"],
       reasoningEfforts: ["minimal", "high"],
     };
-    const wrapper = mount(MainChatModelPopover, {
+    const wrapper = mount(MainChatModelSelectors, {
       props: {
-        ...popoverBaseProps,
+        ...selectorBaseProps,
         agents: [{ id: "custom", name: "Custom", ready: true }],
         activeAgentId: "custom",
         models: [model],
@@ -78,12 +78,12 @@ describe("MainChat model selector", () => {
     wrapper.unmount();
   });
 
-  it("opens one header popover for model and reasoning changes", async () => {
+  it("exposes independent native dropdowns for model and reasoning changes", async () => {
     const model = makeModel("gpt-5.6", "GPT-5.6", "openai");
     model.configJson = { reasoningEfforts: ["low", "high"] };
-    const wrapper = mount(MainChatModelPopover, {
+    const wrapper = mount(MainChatModelSelectors, {
       props: {
-        ...popoverBaseProps,
+        ...selectorBaseProps,
         agents: [{ id: "codex", name: "Codex", ready: true }],
         activeAgentId: "codex",
         models: [model, makeModel("gpt-4.1", "GPT-4.1", "openai")],
@@ -92,23 +92,27 @@ describe("MainChat model selector", () => {
       },
     });
 
-    const toggle = wrapper.get('[data-testid="chat-model-popover-toggle"]');
-    expect(wrapper.find('[data-testid="chat-model-popover-menu"]').attributes("aria-hidden")).toBe("true");
-    await toggle.trigger("click");
-    expect(wrapper.find('[data-testid="chat-model-popover-menu"]').attributes("aria-hidden")).toBe("false");
+    const modelSelect = wrapper.get('select[aria-label="Model"]');
+    const effortSelect = wrapper.get('select[aria-label="Reasoning effort"]');
+    expect(wrapper.findAll("select")).toHaveLength(2);
+    expect(wrapper.find('[role="dialog"]').exists()).toBe(false);
+    expect((modelSelect.element as HTMLSelectElement).value).toBe("gpt-5.6");
+    expect((effortSelect.element as HTMLSelectElement).value).toBe("low");
 
-    await wrapper.find('[data-testid="chat-model-option"][data-model-id="gpt-4.1"]').trigger("click");
+    await modelSelect.setValue("gpt-4.1");
     expect(wrapper.emitted("setModel")?.at(-1)?.[0]).toBe("gpt-4.1");
+    expect(wrapper.emitted("setReasoningEffort")).toBeUndefined();
 
-    await wrapper.find('[data-testid="chat-reasoning-effort"] [data-reasoning-effort="high"]').trigger("click");
+    await effortSelect.setValue("high");
     expect(wrapper.emitted("setReasoningEffort")?.at(-1)?.[0]).toBe("high");
+    expect(wrapper.emitted("setModel")).toHaveLength(1);
     wrapper.unmount();
   });
 
   it("offers the extended Codex reasoning efforts when a model has no override", () => {
-    const wrapper = mount(MainChatModelPopover, {
+    const wrapper = mount(MainChatModelSelectors, {
       props: {
-        ...popoverBaseProps,
+        ...selectorBaseProps,
         agents: [{ id: "codex", name: "Codex", ready: true }],
         activeAgentId: "codex",
         models: [makeModel("gpt-5.6", "GPT-5.6", "openai")],
@@ -131,9 +135,9 @@ describe("MainChat model selector", () => {
     model.configJson = {
       reasoningEfforts: ["medium", "high", "xhigh", "max", "ultra"],
     };
-    const wrapper = mount(MainChatModelPopover, {
+    const wrapper = mount(MainChatModelSelectors, {
       props: {
-        ...popoverBaseProps,
+        ...selectorBaseProps,
         agents: [{ id: "codex", name: "Codex", ready: true }],
         activeAgentId: "codex",
         models: [model],
@@ -143,7 +147,7 @@ describe("MainChat model selector", () => {
     });
 
     const effortSelect = wrapper.find('[data-testid="chat-reasoning-effort"]');
-    expect(effortSelect.find('[aria-pressed="true"]').attributes("data-reasoning-effort")).toBe("ultra");
+    expect((effortSelect.element as HTMLSelectElement).value).toBe("ultra");
     expect(effortSelect.findAll("[data-reasoning-effort]").map((option) => option.attributes("data-reasoning-effort"))).toEqual([
       "medium",
       "high",
@@ -159,9 +163,9 @@ describe("MainChat model selector", () => {
     model.configJson = {
       reasoningEfforts: ["low", "medium", "high", "xhigh", "max"],
     };
-    const wrapper = mount(MainChatModelPopover, {
+    const wrapper = mount(MainChatModelSelectors, {
       props: {
-        ...popoverBaseProps,
+        ...selectorBaseProps,
         agents: [{ id: "claude", name: "Claude", ready: true }],
         activeAgentId: "claude",
         models: [model],
@@ -171,7 +175,7 @@ describe("MainChat model selector", () => {
     });
 
     const effortSelect = wrapper.find('[data-testid="chat-reasoning-effort"]');
-    expect(effortSelect.find('[aria-pressed="true"]').attributes("data-reasoning-effort")).toBe("max");
+    expect((effortSelect.element as HTMLSelectElement).value).toBe("max");
     expect(effortSelect.findAll("[data-reasoning-effort]").map((option) => option.attributes("data-reasoning-effort"))).toEqual([
       "low",
       "medium",
@@ -217,9 +221,9 @@ describe("MainChat model selector", () => {
   });
 
   it("defaults to the first available model when current is unset", async () => {
-    const wrapper = mount(MainChatModelPopover, {
+    const wrapper = mount(MainChatModelSelectors, {
       props: {
-        ...popoverBaseProps,
+        ...selectorBaseProps,
         agents: [{ id: "codex", name: "Codex", ready: true }],
         activeAgentId: "codex",
         models: [makeModel("gpt-4.1", "GPT-4.1", "openai"), makeModel("gpt-4o", "GPT-4o", "openai")],
@@ -233,9 +237,9 @@ describe("MainChat model selector", () => {
   });
 
   it("falls back to the first filtered model for the active agent", async () => {
-    const wrapper = mount(MainChatModelPopover, {
+    const wrapper = mount(MainChatModelSelectors, {
       props: {
-        ...popoverBaseProps,
+        ...selectorBaseProps,
         agents: [
           { id: "claude", name: "Claude", ready: true },
           { id: "codex", name: "Codex", ready: true },
@@ -255,9 +259,9 @@ describe("MainChat model selector", () => {
   });
 
   it("shows only models supported by the active CLI and never switches CLI from model selection", async () => {
-    const wrapper = mount(MainChatModelPopover, {
+    const wrapper = mount(MainChatModelSelectors, {
       props: {
-        ...popoverBaseProps,
+        ...selectorBaseProps,
         agents: [
           { id: "codex", name: "Codex", ready: true },
           { id: "claude", name: "Claude Code", ready: true },
@@ -271,11 +275,11 @@ describe("MainChat model selector", () => {
       },
     });
 
-    const modelSelect = wrapper.find('[data-testid="chat-model-popover-toggle"]');
+    const modelSelect = wrapper.find('[data-testid="chat-model-select"]');
     expect(wrapper.findAll('[data-testid="chat-model-option"]').map((option) => option.attributes("data-model-id"))).toEqual(["gpt-4.1"]);
     expect(modelSelect.text()).not.toContain("Claude Opus 5");
 
-    await wrapper.find('[data-testid="chat-model-option"]').trigger("click");
+    await modelSelect.setValue("gpt-4.1");
 
     expect(wrapper.emitted("switchAgent")).toBeUndefined();
     expect(wrapper.emitted("setModel")?.[0]?.[0]).toBe("gpt-4.1");
@@ -283,9 +287,9 @@ describe("MainChat model selector", () => {
   });
 
   it("does not expose models when no CLI is ready", () => {
-    const wrapper = mount(MainChatModelPopover, {
+    const wrapper = mount(MainChatModelSelectors, {
       props: {
-        ...popoverBaseProps,
+        ...selectorBaseProps,
         agents: [
           { id: "codex", name: "Codex", ready: false },
           { id: "claude", name: "Claude Code", ready: false },
@@ -299,12 +303,13 @@ describe("MainChat model selector", () => {
       },
     });
 
-    const modelSelect = wrapper.find('[data-testid="chat-model-popover-toggle"]');
-    expect(modelSelect.attributes("disabled")).toBeUndefined();
+    const modelSelect = wrapper.find('[data-testid="chat-model-select"]');
+    expect((modelSelect.element as HTMLSelectElement).disabled).toBe(true);
     expect(wrapper.findAll('[data-testid="chat-model-option"]')).toHaveLength(0);
-    expect(wrapper.find('[data-testid="chat-model-popover-menu"]').text()).toContain("No compatible models");
-    expect(wrapper.find('[data-testid="chat-model-popover-menu"]').text()).not.toContain("GPT-4.1");
-    expect(wrapper.find('[data-testid="chat-model-popover-menu"]').text()).not.toContain("Claude Opus 5");
+    expect(modelSelect.text()).toContain("No models");
+    expect(modelSelect.text()).not.toContain("GPT-4.1");
+    expect(modelSelect.text()).not.toContain("Claude Opus 5");
+    expect((wrapper.get('[data-testid="chat-reasoning-effort"]').element as HTMLSelectElement).disabled).toBe(true);
 
     wrapper.unmount();
   });
@@ -313,9 +318,9 @@ describe("MainChat model selector", () => {
     const fable = makeModel("claude-fable-5[1m]", "Claude Fable 5", "anthropic");
     const opus = makeModel("claude-opus-5[1m]", "Claude Opus 5", "anthropic");
     opus.isDefault = true;
-    const wrapper = mount(MainChatModelPopover, {
+    const wrapper = mount(MainChatModelSelectors, {
       props: {
-        ...popoverBaseProps,
+        ...selectorBaseProps,
         agents: [{ id: "claude", name: "Claude Code", ready: true }],
         activeAgentId: "claude",
         models: [fable, opus],
@@ -329,9 +334,9 @@ describe("MainChat model selector", () => {
   });
 
   it("does not emit setModel when the model list is empty", async () => {
-    const wrapper = mount(MainChatModelPopover, {
+    const wrapper = mount(MainChatModelSelectors, {
       props: {
-        ...popoverBaseProps,
+        ...selectorBaseProps,
         agents: [{ id: "codex", name: "Codex", ready: true }],
         activeAgentId: "codex",
         models: [],
@@ -342,10 +347,10 @@ describe("MainChat model selector", () => {
     await wrapper.vm.$nextTick();
     expect(wrapper.emitted("setModel")).toBeUndefined();
 
-    const modelSelect = wrapper.find('[data-testid="chat-model-popover-toggle"]');
+    const modelSelect = wrapper.find('[data-testid="chat-model-select"]');
     expect(modelSelect.exists()).toBe(true);
-    expect(modelSelect.attributes("disabled")).toBeUndefined();
-    expect(wrapper.find('[data-testid="chat-model-popover-menu"]').text()).toContain("No compatible models");
+    expect((modelSelect.element as HTMLSelectElement).disabled).toBe(true);
+    expect(modelSelect.text()).toContain("No models");
 
     wrapper.unmount();
   });
@@ -355,9 +360,9 @@ describe("MainChat model selector", () => {
     model.configJson = {
       reasoningEfforts: ["medium", "high"],
     };
-    const wrapper = mount(MainChatModelPopover, {
+    const wrapper = mount(MainChatModelSelectors, {
       props: {
-        ...popoverBaseProps,
+        ...selectorBaseProps,
         inputLocked: true,
         agents: [{ id: "codex", name: "Codex", ready: true }],
         activeAgentId: "codex",
@@ -374,6 +379,54 @@ describe("MainChat model selector", () => {
     await wrapper.setProps({ inputLocked: false });
     await wrapper.vm.$nextTick();
 
+    expect(wrapper.emitted("setModel")).toBeUndefined();
+    expect(wrapper.emitted("setReasoningEffort")).toBeUndefined();
+    expect((wrapper.get('[data-testid="chat-model-select"]').element as HTMLSelectElement).value).toBe("removed-model");
+    expect((wrapper.get('[data-testid="chat-reasoning-effort"]').element as HTMLSelectElement).disabled).toBe(true);
+    wrapper.unmount();
+  });
+
+  it.each([
+    { busy: true },
+    { inputLocked: true },
+    { connected: false },
+  ])("blocks both dropdown changes when restricted by %j", async (restriction) => {
+    const wrapper = mount(MainChatModelSelectors, {
+      props: {
+        ...selectorBaseProps,
+        ...restriction,
+        agents: [{ id: "codex", name: "Codex", ready: true }],
+        activeAgentId: "codex",
+        models: [makeModel("gpt-4.1", "GPT-4.1", "openai"), makeModel("gpt-4o", "GPT-4o", "openai")],
+        modelId: "gpt-4.1",
+        modelReasoningEffort: "high",
+      },
+    });
+    const modelSelect = wrapper.get('[data-testid="chat-model-select"]');
+    const effortSelect = wrapper.get('[data-testid="chat-reasoning-effort"]');
+    expect((modelSelect.element as HTMLSelectElement).disabled).toBe(true);
+    expect((effortSelect.element as HTMLSelectElement).disabled).toBe(true);
+    await modelSelect.setValue("gpt-4o");
+    await effortSelect.setValue("ultra");
+    expect(wrapper.emitted("setModel")).toBeUndefined();
+    expect(wrapper.emitted("setReasoningEffort")).toBeUndefined();
+    wrapper.unmount();
+  });
+
+  it("updates the selected native values when active-lane props change", async () => {
+    const wrapper = mount(MainChatModelSelectors, {
+      props: {
+        ...selectorBaseProps,
+        agents: [{ id: "codex", name: "Codex", ready: true }],
+        activeAgentId: "codex",
+        models: [makeModel("gpt-4.1", "GPT-4.1", "openai"), makeModel("gpt-4o", "GPT-4o", "openai")],
+        modelId: "gpt-4.1",
+        modelReasoningEffort: "high",
+      },
+    });
+    await wrapper.setProps({ modelId: "gpt-4o", modelReasoningEffort: "ultra" });
+    expect((wrapper.get('[data-testid="chat-model-select"]').element as HTMLSelectElement).value).toBe("gpt-4o");
+    expect((wrapper.get('[data-testid="chat-reasoning-effort"]').element as HTMLSelectElement).value).toBe("ultra");
     expect(wrapper.emitted("setModel")).toBeUndefined();
     expect(wrapper.emitted("setReasoningEffort")).toBeUndefined();
     wrapper.unmount();
